requirements.txt
anyio==4.3.0
certifi==2024.2.2
charset-normalizer==3.3.2
construct==2.10.68
construct-typing==0.5.6
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
jsonalias==0.1.1
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
sniffio==1.3.1
solana==0.34.0
solders==0.21.0
typing_extensions==4.11.0
urllib3==2.2.1
websockets==11.0.3